+++ tfhe/src/boolean/parameters/mod.rs
@@ -24,7 +24,8 @@ pub use crate::core_crypto::commons::parameters::{
     LweDimension, PolynomialSize,
 };
 
-use crate::shortint::keycache::NamedParam;
+#[cfg(any(test, doctest, feature = "internal-keycache"))]
+use crate::keycache::NamedParam;
 use serde::{Deserialize, Serialize};
 
 /// A set of cryptographic parameters for homomorphic Boolean circuit evaluation.
@@ -197,6 +198,7 @@ pub const TFHE_LIB_PARAMETERS: BooleanParameters = BooleanParameters {
 
 pub const VEC_BOOLEAN_PARAM: [BooleanParameters; 2] = [DEFAULT_PARAMETERS, TFHE_LIB_PARAMETERS];
 
+#[cfg(any(test, doctest, feature = "internal-keycache"))]
 impl NamedParam for BooleanParameters {
     fn name(&self) -> &'static str {
         if *self == DEFAULT_PARAMETERS {
